The first upload!
In this batch of code, I have included the C++ code I wrote to do kinematics calculations for the robot, as well as the API we were provided that actually controls the robot (a LabVolt 5150). I have to do a bit more digging to find the code I wrote that actually controls the robot.